cleanup: helper

+++ Iris/Classes/Helper.cs
@@ -1,4 +1,3 @@
-using SFML.Window;
 using SFML.System;
 using SFML.Graphics;
 using System;
@@ -12,19 +11,9 @@ namespace Iris
     {
         public static Vector2f PolarToVector2(float radius, float angle, float xPercent, float yPercent)
         {
-            return new Vector2f(radius * (float)Math.Cos((double)angle) * xPercent, radius * (float)Math.Sin((double)angle) * yPercent);
+            return new Vector2f(radius * (float)Math.Cos(angle) * xPercent,
+                                radius * (float)Math.Sin(angle) * yPercent);
         }
-
-        //public static Vector2f Lerp(Vector2f point1, Vector2f point2, float percent)
-        //{
-        //    Vector2f lerpPos = new Vector2f();
-        //    lerpPos.X = (point2.X - point1.X) * percent;
-        //}
-
-        //public static Vector2f lerp(Vector2f A, Vector2f B, float t )
-        //{
-        //    return A * t + B*(1f - t);
-        //}
 
         public static Vector2f lerp(Vector2f value1, Vector2f value2, float amount)
         {
